Got UV meter working on Mac and ESP
This is very rough. It shows the connecting strings and then shows the time and UV values, updating once a second. Need to do better layout and add support for a button

The display is wonky. It only works every OTHER reset. But it always works on the first power on. Since I won't be using reset in practice, this will be fine.