Defer calling display() until the main thread
This was causing a crash when display was called from a timer firing, so now display() is called in the main loop. But now some messages don't display because they are sent before the main loop starts (network messages). I think I will try the other Ticker library which calls its callbacks in the main loop.